+++ nodes/Autotask/resources/companySiteConfigurations/index.ts
@@ -1,0 +1,2 @@
+export { companySiteConfigurationFields } from './description';
+export { executeCompanySiteConfigurationOperation } from './execute';